- fix cocos api interface

+++ api/jsb_cocos2dx_baseframework_auto_api.js
@@ -162,7 +162,7 @@ getError : function (
 getLong : function (
 )
 {
-    return long long;
+    return long;
 },
 
 /**
